+++ packages/react-ui/src/components/EventForm/EventForm.tsx
@@ -1,5 +1,6 @@
 import {LoadingButton} from '@mui/lab';
 import {
+  Alert,
   Box,
   BoxProps,
   FormControl,
@@ -91,7 +92,8 @@ const classes = {
   actions: `${PREFIX}-actions`,
   privacySection: `${PREFIX}-privacy-section`,
   privacySectionInfo: `${PREFIX}-privacy-section-info`,
-  error: `${PREFIX}-error`
+  error: `${PREFIX}-error`,
+  genericError: `${PREFIX}-generic-error`
 };
 
 const Root = styled(Box, {
@@ -199,6 +201,7 @@ export default function EventForm(inProps: EventFormProps): JSX.Element {
   // STATE
   const [field, setField] = useState<InitialFieldState>(initialFieldState);
   const [error, setError] = useState<any>({});
+  const [genericError, setGenericError] = useState<string | null>(null);
 
   // PREFERENCES
   const scPreferences: SCPreferencesContextType = useSCPreferences();
@@ -234,6 +237,7 @@ export default function EventForm(inProps: EventFormProps): JSX.Element {
 
         setError(error);
       }
+      setGenericError(null);
     },
     [error]
   );
@@ -260,11 +264,12 @@ export default function EventForm(inProps: EventFormProps): JSX.Element {
       ...prev,
       ...data
     }));
+    setGenericError(null);
   }, []);
 
   const handleSubmit = useCallback(() => {
     setField((prev) => ({...prev, ['isSubmitting']: true}));
-
+    setGenericError(null);
     const formData = new FormData();
 
     if (field.imageOriginalFile) {
@@ -310,16 +315,30 @@ export default function EventForm(inProps: EventFormProps): JSX.Element {
       })
       .catch((e) => {
         const _error = formatHttpErrorCode(e);
-
-        if (Object.values(_error)[0]['error'] === 'unique') {
-          setError({
-            ...error,
-            ['nameError']: <FormattedMessage id="ui.eventForm.name.error.unique" defaultMessage="ui.eventForm.name.error.unique" />
-          });
+        if ('errorsError' in _error || !Object.keys(_error).length) {
+          setGenericError(
+            intl.formatMessage({
+              id: 'ui.eventForm.genericError',
+              defaultMessage: 'ui.eventForm.genericError'
+            })
+          );
         } else {
-          setError({...error, ..._error});
+          setGenericError(null);
         }
-
+        let __errors = {};
+        if ('coverError' in _error) {
+          __errors = {
+            ...__errors,
+            ['coverError']: <FormattedMessage id="ui.ui.eventForm.cover.error" defaultMessage="ui.ui.eventForm.cover.error" />
+          };
+        }
+        if ('nameError' in _error || ('nonFieldErrorsError' in _error && _error['nonFieldErrorsError'].error === 'unique')) {
+          __errors = {
+            ...__errors,
+            ['nameError']: <FormattedMessage id="ui.eventForm.name.error.unique" defaultMessage="ui.eventForm.name.error.unique" />
+          };
+        }
+        setError(__errors);
         setField((prev) => ({...prev, ['isSubmitting']: false}));
         Logger.error(SCOPE_SC_UI, e);
         onError?.(e);
@@ -333,9 +352,9 @@ export default function EventForm(inProps: EventFormProps): JSX.Element {
 
       if (error[`${name}Error`]) {
         delete error[`${name}Error`];
-
         setError(error);
       }
+      setGenericError(null);
     },
     [setField, error]
   );
@@ -346,13 +365,12 @@ export default function EventForm(inProps: EventFormProps): JSX.Element {
 
       if (error[`${name}Error`]) {
         delete error[`${name}Error`];
-
         setError(error);
       } else if (error['endDateError']) {
         delete error['endDateError'];
-
         setError(error);
       }
+      setGenericError(null);
     },
     [setField, error]
   );
@@ -391,6 +409,7 @@ export default function EventForm(inProps: EventFormProps): JSX.Element {
       <Paper style={_backgroundCover} classes={{root: classes.cover}}>
         <UploadEventCover isCreationMode={true} onChange={handleChangeCover} />
       </Paper>
+      {Boolean(error['coverError']) && <Typography color="error">{error['coverError']}</Typography>}
       <FormGroup className={classes.form}>
         <TextField
           required
@@ -657,6 +676,13 @@ export default function EventForm(inProps: EventFormProps): JSX.Element {
             ) : null
           }
         />
+        {genericError && (
+          <Box className={classes.genericError}>
+            <Alert variant="filled" severity="error">
+              {genericError}
+            </Alert>
+          </Box>
+        )}
         <Box className={classes.actions}>
           <LoadingButton
             loading={field.isSubmitting}
@@ -669,7 +695,7 @@ export default function EventForm(inProps: EventFormProps): JSX.Element {
               (field.location === SCEventLocationType.ONLINE && !field.link) ||
               (field.location === SCEventLocationType.PERSON && !field.geolocation) ||
               (field.recurring !== SCEventRecurrenceType.NEVER && !field.endDate && !field.endTime) ||
-              Object.keys(error).length !== 0 ||
+              field.isSubmitting ||
               field.name.length > EVENT_TITLE_MAX_LENGTH ||
               field.description.length > EVENT_DESCRIPTION_MAX_LENGTH
             }